+++ packages/cli/src/metadata.ts
@@ -4,8 +4,8 @@ import * as fs from "node:fs/promises"
 import {
   HexString,
   metadata,
-  NormalizedMetadata,
-  normalizeMetadata,
+  UnifiedMetadata,
+  unifyMetadata,
   v15,
 } from "@polkadot-api/substrate-bindings"
 import { getWsProvider } from "@polkadot-api/ws-provider/node"
@@ -119,7 +119,7 @@ const getMetadataFromWsURL = async (wsURL: string) =>
   getMetadataCall(withPolkadotSdkCompat(getWsProvider(wsURL)))
 
 export async function getMetadata(entry: EntryConfig): Promise<{
-  metadata: NormalizedMetadata
+  metadata: UnifiedMetadata
   metadataRaw: Uint8Array
   genesis?: HexString
 } | null> {
@@ -129,11 +129,11 @@ export async function getMetadata(entry: EntryConfig): Promise<{
     const data = await fs.readFile(entry.metadata)
     const metadataRaw = new Uint8Array(data)
 
-    let meta: NormalizedMetadata
+    let meta: UnifiedMetadata
     try {
-      meta = normalizeMetadata(metadata.dec(metadataRaw))
+      meta = unifyMetadata(metadata.dec(metadataRaw))
     } catch (_) {
-      meta = normalizeMetadata(v15.dec(metadataRaw))
+      meta = unifyMetadata(v15.dec(metadataRaw))
     }
 
     return {