+++ packages/signers/meta-signers/src/get-codecs.ts
@@ -1,11 +1,8 @@
 import { getDynamicBuilder, getLookupFn } from "@polkadot-api/metadata-builders"
-import {
-  decAnyMetadata,
-  normalizeMetadata,
-} from "@polkadot-api/substrate-bindings"
+import { decAnyMetadata, unifyMetadata } from "@polkadot-api/substrate-bindings"
 
 export function getCodecs(metadata: Uint8Array) {
-  const tmpMeta = normalizeMetadata(decAnyMetadata(metadata))
+  const tmpMeta = unifyMetadata(decAnyMetadata(metadata))
   const lookup = getLookupFn(tmpMeta)
   if (lookup.call === null) throw new Error("Unsupported metadata")
   const dynamicBuilder = getDynamicBuilder(lookup)